Extract EventRepeats as functional component
